feat(UNIVERSAL_SESSION_STORAGE): add SESSION_STORAGE support
