+++ projects/universal/package.json
@@ -3,7 +3,7 @@
     "version": "1.5.0",
     "peerDependencies": {
         "@angular/core": ">=6.0.0",
-        "@ng-web-apis/common": "1.4.0",
+        "@ng-web-apis/common": "1.6.0",
         "@types/node": ">=10.0.0",
         "rxjs": ">=6.0.0"
     },
